ESDK-1676 handle exceptions being unchecked in Kotlin
Since in Kotlin all exceptions are unchecked, the unit test which mocks the resolveFromOpenStreetMaps method and simulates an IOException fails, since this is a checked exception. To make this work with PowerMock the Throws annotation is needed in the Kotlin implementation. This will be translated into byte code the same way as a Java throws keyword on methods.

More on Exceptions in Kotlin:
https://medium.com/google-developer-experts/those-sneaky-kotlin-exceptions-597c93d29358